Добавлен webpack. Разбита работа js на компоненты.
