+++ src/components/film-more-btn.js
@@ -1,0 +1,3 @@
+export const createFilmMoreButton = () => {
+  return `<button class="films-list__show-more">Show more</button>`;
+};
