+++ src/components/film-statistics.js
@@ -1,0 +1,7 @@
+export const createFilmStatistics = () => {
+  return (
+    `<section class="footer__statistics">
+        <p>130 291 movies inside</p>
+    </section>`
+  );
+};
